refactor: 💡 extract
